Update webpage router

+++ runner/index.es6.js
@@ -3,6 +3,7 @@ import co from 'co';
 
 const WEBSERVER_PORT = process.env.WB_WSERVER_PORT;
 const INTERNAL_WEBSERVER_PORT = process.env.WB_INTN_WSERVER_PORT;
+const INIT = process.env.WEBLOG_ENV === 'init';
 
 WeblogJS.init({
     webPort: WEBSERVER_PORT,
@@ -18,13 +19,26 @@ const admin = Object.freeze({
 });
 
 co(function* () {
-    yield WeblogJS.webServer.start();
-    yield WeblogJS.internalWebServer.start();
-    yield WeblogJS.dbSettingOperator.dropDatabase();
-    yield WeblogJS.dbSettingOperator.createIndexes();
-    yield WeblogJS.dbSettingOperator.removeAllDocuments();
-    yield WeblogJS.createUser(admin);
-    console.log("Web Server has started...");
+    if (INIT) {
+        console.log("Dropping the database...");
+        yield WeblogJS.dbSettingOperator.dropDatabase();
+
+        console.log("Creating indexes for collections...");
+        yield WeblogJS.dbSettingOperator.createIndexes();
+
+        console.log("Removing all the documents in the database...");
+        yield WeblogJS.dbSettingOperator.removeAllDocuments();
+
+        console.log("Creating the admin account...");
+        yield WeblogJS.createUser(admin);
+
+        console.log("Finished the initialization.");
+        process.exit();
+    } else {
+        yield WeblogJS.webServer.start();
+        yield WeblogJS.internalWebServer.start();
+        console.log("Web Server has started...");
+    }
 }).catch((error) => {
     console.error(error);
 });
